Refactored: Converted to control
